feat(app): add design system widgets, screens, router and reference data layer
Screens: Dashboard, Nutrition, Add Food, Train, and Live Gym Mode — the last
built to the one-handed, offline-first constraints in docs/06 (64 dp+ targets
in the thumb zone, local commit before anything else, prefill from the last
performance).

Data layer ships as reference in-memory implementations of the full repository
contracts. They model the production write ordering — commit locally, replicate
after — so swapping in Firestore/Drift is two provider overrides and nothing
above the interfaces moves.

Also adds tool/check_sources.py, which enforces the two rules the analyzer
cannot: the pure-Dart domain boundary, and no colour literal outside
core/theme. Both pass.

Fixes found while writing it: several num-vs-double assignment errors from
clamp()/math.max() returning num; state fields mutated during build in Live
Gym; and colour literals that had escaped the theme.

+++ app/lib/core/engines/macro_calculator.dart
@@ -267,7 +267,7 @@ abstract final class MacroCalculator {
     // macronutrient the goal depends on.
     if (carbKcal < 0) {
       fatG = math.max(0.5 * weightKg, (kcal - proteinKcal) / 9 * 0.9);
-      carbKcal = math.max(0, kcal - proteinKcal - fatG * 9);
+      carbKcal = math.max(0.0, kcal - proteinKcal - fatG * 9);
     }
 
     return MacroTargets(
@@ -294,7 +294,7 @@ abstract final class MacroCalculator {
     required double tdeeRest,
     required int trainingDays,
   }) {
-    final t = trainingDays.clamp(0, 7);
+    final t = trainingDays.clamp(0, 7).toInt();
     final r = 7 - t;
     return (trainingKcal - tdeeTraining) * t + (restKcal - tdeeRest) * r;
   }
